de-conflate LaunchAgent with prefs
The Plex LaunchAgent lives in Library/LaunchAgents and is custom-written in plex-setup to launch Plex.
The Plex preferences live in Library/Preferences and is maintained by the Plex service, with seeded settings by plex-setup.
These should not have the same name, and their names should be defined at the top of the script, not hardcoded in various places throughout.